Refactored SQL queries to LINQ
Refactored the code to use LINQ for data querying instead of raw SQL queries. The `System.Linq` namespace has been added to the using directives. The code for executing the SQL query, reading the results, and disposing the reader has been replaced with a LINQ query that groups students by their enrollment date and counts the number of students in each group. The results of the LINQ query are added to the `groups` list. The code for getting the database schema and checking the operating system, as well as the `dbSchema` variable and related code, have been commented out.